[Testcontainers] move integration test to internal (#855)
* feat: use testcontainers-go for the integration tests

* chore: bump to latest release

* chore: wrap the ES container into a struct

* adapt github action to testcontainers

* use the same container for all the integration subtests

* use testcontainers for esapi integration tests

* use testcontainers for esutil integration tests

* update testcontainers-go dependency to main branch as a wip

* move custom transport test and insecure default to insecure integration part

* update to latest testcontainers-go 0.31.0

* reorder imports

* move integration test to internal package with testcontainers

* adapt makefile to use the new internal/testing package

* fix bad version transport regression

* update examples dependencies

* reinstate server launch for example testing

---------
