+++ _examples/extension/go.mod
@@ -1,0 +1,16 @@
+module github.com/elastic/go-elasticsearch/v8/_examples/extension
+
+go 1.21
+
+require (
+	github.com/elastic/elastic-transport-go/v8 v8.5.0
+	github.com/elastic/go-elasticsearch/v8 v8.13.1
+)
+
+require (
+	github.com/go-logr/logr v1.3.0 // indirect
+	github.com/go-logr/stdr v1.2.2 // indirect
+	go.opentelemetry.io/otel v1.21.0 // indirect
+	go.opentelemetry.io/otel/metric v1.21.0 // indirect
+	go.opentelemetry.io/otel/trace v1.21.0 // indirect
+)
