+++ internal/testing/e2e/esapi_integration_test.go
@@ -18,14 +18,17 @@
 //go:build integration
 // +build integration
 
-package esapi_test
+package e2e_test
 
 import (
 	"bytes"
+	"context"
 	"encoding/json"
 	"fmt"
+	"os"
 	"strings"
 	"testing"
+	"testing/containertest"
 	"time"
 
 	"github.com/elastic/go-elasticsearch/v8"
@@ -33,8 +36,25 @@ import (
 )
 
 func TestAPI(t *testing.T) {
+	stackVersion := elasticsearch.Version
+	if v := os.Getenv("STACK_VERSION"); v != "" {
+		stackVersion = v
+	}
+
+	elasticsearchSrv, err := containertest.NewElasticsearchService(stackVersion)
+	if err != nil {
+		t.Fatalf("Error setting up Elasticsearch container: %s", err)
+	}
+	defer func() {
+		if err := elasticsearchSrv.Terminate(context.Background()); err != nil {
+			t.Fatalf("Error terminating Elasticsearch container: %s", err)
+		}
+	}()
+
+	tcCfg := elasticsearchSrv.ESConfig()
+
 	t.Run("Search", func(t *testing.T) {
-		es, err := elasticsearch.NewDefaultClient()
+		es, err := elasticsearch.NewClient(tcCfg)
 		if err != nil {
 			t.Fatalf("Error creating the client: %s\n", err)
 		}
@@ -59,7 +79,7 @@ func TestAPI(t *testing.T) {
 	})
 
 	t.Run("Headers", func(t *testing.T) {
-		es, err := elasticsearch.NewDefaultClient()
+		es, err := elasticsearch.NewClient(tcCfg)
 		if err != nil {
 			t.Fatalf("Error creating the client: %s\n", err)
 		}
@@ -89,7 +109,7 @@ func TestAPI(t *testing.T) {
 			requestID = "reindex-123"
 		)
 
-		es, err := elasticsearch.NewDefaultClient()
+		es, err := elasticsearch.NewClient(tcCfg)
 		if err != nil {
 			t.Fatalf("Error creating the client: %s\n", err)
 		}
