+++ internal/testing/e2e/json_reader_integration_test.go
@@ -18,25 +18,45 @@
 //go:build integration
 // +build integration
 
-package esutil_test
+package e2e_test
 
 import (
-	"strings"
-	"testing"
-
+	"context"
 	"github.com/elastic/go-elasticsearch/v8"
 	"github.com/elastic/go-elasticsearch/v8/esapi"
 	"github.com/elastic/go-elasticsearch/v8/esutil"
+	"os"
+	"strings"
+	"testing"
+
+	"testing/containertest"
 )
 
 func TestJSONReaderIntegration(t *testing.T) {
+	stackVersion := elasticsearch.Version
+	if v := os.Getenv("STACK_VERSION"); v != "" {
+		stackVersion = v
+	}
+
+	elasticsearchSrv, err := containertest.NewElasticsearchService(stackVersion)
+	if err != nil {
+		t.Fatalf("Error setting up Elasticsearch container: %s", err)
+	}
+	defer func() {
+		if err := elasticsearchSrv.Terminate(context.Background()); err != nil {
+			t.Fatalf("Error terminating Elasticsearch container: %s", err)
+		}
+	}()
+
+	tcCfg := elasticsearchSrv.ESConfig()
+
 	t.Run("Index and search", func(t *testing.T) {
 		var (
 			res *esapi.Response
 			err error
 		)
 
-		es, err := elasticsearch.NewDefaultClient()
+		es, err := elasticsearch.NewClient(tcCfg)
 		if err != nil {
 			t.Fatalf("Error creating the client: %s\n", err)
 		}
